Extracting creating comparator to SortHelper class.
