+++ components/writers/src/main/java/org/datacleaner/extension/output/CreateCsvFileAnalyzer.java
@@ -38,7 +38,6 @@ import org.apache.metamodel.data.Row;
 import org.apache.metamodel.schema.Table;
 import org.apache.metamodel.util.FileHelper;
 import org.apache.metamodel.util.FileResource;
-import org.apache.metamodel.util.NumberComparator;
 import org.apache.metamodel.util.Resource;
 import org.datacleaner.api.Alias;
 import org.datacleaner.api.Categorized;
@@ -55,8 +54,6 @@ import org.datacleaner.api.Validate;
 import org.datacleaner.beans.writers.WriteDataResult;
 import org.datacleaner.beans.writers.WriteDataResultImpl;
 import org.datacleaner.components.categories.WriteSuperCategory;
-import org.datacleaner.components.convert.ConvertToDateTransformer;
-import org.datacleaner.components.convert.ConvertToNumberTransformer;
 import org.datacleaner.connection.CsvDatastore;
 import org.datacleaner.connection.Datastore;
 import org.datacleaner.descriptors.FilterDescriptor;
@@ -65,8 +62,6 @@ import org.datacleaner.job.builder.AnalysisJobBuilder;
 import org.datacleaner.output.OutputWriter;
 import org.datacleaner.output.csv.CsvOutputWriterFactory;
 import org.datacleaner.user.UserPreferences;
-import org.datacleaner.util.CompareUtils;
-import org.datacleaner.util.ReflectionUtils;
 import org.datacleaner.util.sort.SortMergeWriter;
 
 @Named("Create CSV file")
@@ -205,7 +200,7 @@ public class CreateCsvFileAnalyzer extends AbstractOutputWriterAnalyzer implemen
             final CsvDataContext tempDataContext = new CsvDataContext(_targetFile, csvConfiguration);
             final Table table = tempDataContext.getDefaultSchema().getTable(0);
 
-            final Comparator<? super Row> comparator = createComparator();
+            final Comparator<? super Row> comparator = SortHelper.createComparator(columnToBeSortedOn, indexOfColumnToBeSortedOn);
 
             final CsvWriter csvWriter = new CsvWriter(csvConfiguration);
             final SortMergeWriter<Row, Writer> sortMergeWriter = new SortMergeWriter<Row, Writer>(comparator) {
@@ -267,58 +262,6 @@ public class CreateCsvFileAnalyzer extends AbstractOutputWriterAnalyzer implemen
         final Resource resource = new FileResource(file);
         final Datastore datastore = new CsvDatastore(file.getName(), resource, csvConfiguration);
         return new WriteDataResultImpl(rowCount, datastore, null, null);
-    }
-
-    private Comparator<Row> createComparator() {
-        final Class<?> dataType = columnToBeSortedOn.getDataType();
-        final boolean isNumber = dataType != null && ReflectionUtils.isNumber(dataType);
-        final boolean isDate = dataType != null && ReflectionUtils.isDate(dataType);
-
-        return new Comparator<Row>() {
-            @Override
-            public int compare(Row row1, Row row2) {
-                final Comparable<?> value1 = getComparableValue(row1, isNumber, isDate);
-                final Comparable<?> value2 = getComparableValue(row2, isNumber, isDate);
-
-                final int comparableResult = CompareUtils.compareUnbound(value1, value2);
-                if (comparableResult != 0) {
-                    return comparableResult;
-                } else {
-                    // The values of the data at the row, and column to be
-                    // sorted on are
-                    // exactly the same. Now look at other values of all the
-                    // columns to
-                    // find if the two rows are same.
-                    int numberOfSelectItems = row1.getSelectItems().length;
-                    for (int i = 0; i < numberOfSelectItems; i++) {
-                        final String rowValue1 = (String) row1.getValue(i);
-                        final String rowValue2 = (String) row2.getValue(i);
-                        if (CompareUtils.compare(rowValue1, rowValue2) == 0) {
-                            continue;
-                        } else {
-                            return CompareUtils.compare(rowValue1, rowValue2);
-                        }
-                    }
-                }
-
-                return comparableResult;
-            }
-        };
-    }
-
-    protected Comparable<?> getComparableValue(Row row, boolean isNumber, boolean isDate) {
-        final String value = (String) row.getValue(indexOfColumnToBeSortedOn);
-        if (isNumber) {
-            final Number result = ConvertToNumberTransformer.transformValue(value);
-            if (result instanceof Comparable) {
-                return (Comparable<?>) result;
-            }
-            return NumberComparator.getComparable(result);
-        }
-        if (isDate) {
-            return ConvertToDateTransformer.getInternalInstance().transformValue(value);
-        }
-        return value;
     }
 
     public void setFile(File file) {
